반응별랭킹 탑텐 요청 타입 추가
