+++ client/web/src/organisms/mainpage/ranking/topten/TrendsBarChart.tsx
@@ -2,7 +2,8 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useTheme } from '@material-ui/core/styles';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
-import { Scores, TrendsBarChartProps } from '../types/ToptenCard.types';
+import { Scores } from '@truepoint/shared/dist/res/RankingsResTypes.interface';
+import { TrendsBarChartProps } from '../types/ToptenCard.types';
 
 function getSeriesName(currentScoreName: keyof Scores) {
   switch (currentScoreName) {
